web-edge/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
google-cloud-logging==3.10.0